int_user_info/serializers/user_info.py
from rest_framework import serializers
from ..models import UserInfo


class UserInfoSerializer(serializers.ModelSerializer):

    class Meta:
        model = UserInfo
        fields = ('score',)
