requirements.txt
Django==1.9
django-filter==0.11.0
djangorestframework==3.3.2
factory-boy==2.6.0
fake-factory==0.5.3
psycopg2==2.6.1